feat: update demo
