--- packages/site/examples/tool/edgeBundling/API.zh.md
@@ -1,5 +1,0 @@
----
-title: API
----
-
-<embed src="@/docs/api/Plugins.zh.md"></embed>
