--- packages/site/examples/tool/edgeBundling/demo/meta.json
@@ -1,16 +1,0 @@
-{
-  "title": {
-    "zh": "中文分类",
-    "en": "Category"
-  },
-  "demos": [
-    {
-      "filename": "edgeBundling.js",
-      "title": {
-        "zh": "边绑定",
-        "en": "Edge Bundling"
-      },
-      "screenshot": "https://gw.alipayobjects.com/mdn/rms_f8c6a0/afts/img/A*7g4iRqVfkHQAAAAAAAAAAABkARQnAQ"
-    }
-  ]
-}
