--- packages/site/examples/tool/edgeBundling/index.en.md
@@ -1,10 +1,0 @@
----
-title: Edge Bundling
-order: 3
----
-
-Bundling is a built-in components in G6. It is used for bundling the complex edges to reduce the visual clutter.
-
-## Usage
-
-The demo below show how to use Bundling on graph.
