--- packages/site/examples/tool/edgeBundling/index.zh.md
@@ -1,10 +1,0 @@
----
-title: 边绑定
-order: 3
----
-
-G6 中内置的 Bundling 组件，用于边的绑定。
-
-## 使用指南
-
-下面的代码演示展示了如何在图上使用 Bundling。
