--- packages/site/examples/tool/edgeFilterLens/API.en.md
@@ -1,5 +1,0 @@
----
-title: API
----
-
-<embed src="@/docs/api/Plugins.en.md"></embed>
