--- packages/site/examples/tool/edgeFilterLens/index.en.md
@@ -1,10 +1,0 @@
----
-title: Edge Filter Lens
-order: 7
----
-
-Edge Filter Lens is a built-in components for edge filtering in G6.
-
-## Usage
-
-Edge Filter Lens is designed for edge filtering, the desired edges will be kept inside the lens while the others will be hidden. The `type` and `shouldShow` can be assigned to design the conditions. And the lens's style can be configured by `delegateStyle`. The demo below shows how to use fisheye on graph. The API can be found at [Edge Filter Lens Plugin](/en/docs/api/Plugins#edge-filter-lens).
