--- packages/site/examples/tool/edgeFilterLens/index.zh.md
@@ -1,10 +1,0 @@
----
-title: 边过滤镜
-order: 7
----
-
-G6 中内置的边过滤镜组件。
-
-## 使用指南
-
-边过滤镜可以将关注的边保留在过滤镜范围内，其他边将在该范围内不显示。过滤的条件可以通过 `type` 和 `shouldShow` 指定。过滤镜的样式可以通过 `delegateStyle` 指定。下面 demo 演示了如何使用边过滤镜。API 参见 [边过滤镜插件](/zh/docs/api/Plugins#edge-filter-lens)。
